Use the normalized path for unmounting, patch by Peter Daum
In "unmount" it tries to make sure the path to umount is actually mounted
by comparing the mounted path against the output of "/proc/mounts". Because
paths reported by "/proc/mounts" are normalized, this will fail if there
are any symlinks involved (leaving all mounts and an unbootable usb stick).