bsd4grml was not copied correctly to the target device.
This fixes a bug introduced in commit c70267a9. The problem is that the
bsd4grml addon is a directory and it needs a trailing slash for copying
with rysnc. handle_addon_copy just search for the file and copies it
directly to the target device without adding a slash.

The current fix is to remove the explicit destination filename in
handle_addon_copy.

Thanks to  Alex Mestiashvili for reporting that bug.