Fix FORMAT definition for logging module.
Level does not specify that only messages of this level are printed
out, but the minimum level. So a definition like this

    FORMAT = "Debug: %(asctime)-15s %(message)s"
    FORMAT = "%(asctime)-15s %(message)s"

misleads the user into thinking that every message printed out is
actually a debug message, which is not true.
We now use "%(message)s" for everything except verbose mode which
annotates the timestamp to the message.