Use lowercase 'timeout' command in syslinux configuration [Closes: issue1262]
With an uppercase 'TIMEOUT' command the system seems to be failing
to autoboot for a user (using kernel version 3.9-1-amd64 on a
ThinkPad T410s). While syslinux shouldn't care about it
("Keywords are case insensitive") changing the uppercase command
to lowercase fixes the issue, so let's go for it.

Thanks: Martin Šlouf for reporting this bug including a workaround