grml2iso: disable UEFI mode by default, only enable for xorriso under according prerequisites
Getting an error message like:

| isohybrid version supports --uefi option, enabling
| Creating dd-able ISO using isohybrid
| isohybrid: custom_grml.iso: unable to find efi image

is quite confusing, especially if the /efi/boot/bootx64.efi
and /boot/efi.img files are actually present on the image.

When using xorriso everything is fine, but when using genisoimage
within grml2iso the ISO lacks according UEFI support and the
above mention error message appears. So instead only enable UEFI
mode if building with xorriso and all the other prerequisites are
fulfilled as well.