Fixed manpages
Manpages were including the text .5v at the end of some paragraphs. This was
caused by an old hack in the Makefile which removed the text .sp, which was
accidently added to the end of some paragraphs. (sed -i 's/\.sp//') Looks like
this hack is not needed any longer so it was just commented out and so the .5v
disappeared (which might have been the text ".sp 5v")