Support netboot.xyz.lkrn addon
For the grml96 flavours we use grml2iso, which itself
relies on grml2usb for copying the relevant files.
With the addition of the netboot.xyz.lkrn file we
forgot

Actually it would be nice to copy any present addon
from /boot/addons/, but close to the new stable Grml
release let's keep the change as small as possible.

Thanks: Michael Schierl for reporting the bug
Closes grml/grml#25