Github action: do not install virtualenv + python3-setuptools
pip3 and python3-setuptools are available out of the box on
ubuntu-latest.  By not having to install python3-virtualenv we gain ~6
seconds in each of of the codecheck and unit tests runs. So instead
let's rely on system wide pip3, given that we're running in a throw-away
VM anyway.

Thanks: Chris Hofstaedtler