Make blockdev --rereadpt optional
In GitHub Actions I've observed blockdev --rereadpt failing.
The reasons are unknown, but it also appears to be a harmless
failure.

I believe the rereadpt call should be fine to ignore; afterall
what we do is installing an MBR onto a device with a pre-existing
partition table; the partition layout should not change.