Introduce automatic bootloader selection
Introduce a new --bootloader parameter, defaulting to "auto".
It also accepts "syslinux", "grub" and "efi". With "auto", on arm64 it will
choose "efi" (EFI-boot only). On amd64, the first available option, depending
on installed programs, will be chosen.

This also makes syslinux config files optional, and if they are absent
syslinux configuration will be skipped.

The --syslinux and --grub options are effectively deprecated.